(#33) Experimental command plugins
This introduce an experimental way to extend one fisk
command with another. Fisk commands will have --fisk-introspection
that emits an ApplicationModel.

Other fisk apps can import this using the ExternalPluginJSON()
function. Help output and more will include the full command
model and on execution it will call the command to do
the actual work

Ideally plugins will be able to define global flags that clash
with the command they plug into, in that case the value of the
global flag will be passed to the app but for now that's left
for a followup PR

Signed-off-by: R.I.Pienaar <rip@devco.net>