Attempt to improve default true negatable booleans
This kind of works, only shows --no-x rathr than --[no-]-x for
default true negatable booleans, since the --x is implied.

But this has a drawback that help is now pretty weird by default:

 --no-allow-direct                     Allows fast, direct, access to stream data via the direct get API

Here the help is written from the perspective of the --allow-direct but
with this change it now reads weirdly.

Ultimately not a change we can make as it would negatively impact existing apps

Signed-off-by: R.I.Pienaar <rip@devco.net>